+++ docs/changelog.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 **Bug fixes**
 
+* Do not allowed a user to select a signage that is not in it's structure in multi-update form
 * Map screenshot contains wrong layers by default
 
 
@@ -24,6 +25,9 @@ CHANGELOG
 
 * Fix ``label_en`` content on sensitivity module parser
 * Fix restricted area geojson that are not different by type in cache (#5136)
+
+**New features**
+* Add bulk deletion/edition on list views (refs #5107).
 
 
 2.121.1         (2025-11-17)